chore: set transport to wss
